requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.14.0
fsspec==2024.5.0
huggingface-hub==0.23.3
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
lightning-utilities==0.11.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
preshed==3.0.9
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.13.1
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
torchmetrics==1.4.0.post0
tqdm==4.66.4
transformers==4.41.2
typer==0.9.4
typing_extensions==4.12.0
tzdata==2024.1
umlsparser @ git+https://github.com/DATEXIS/UMLSParser.git@3972c3f43db3e9d63e1282ad6e91b8fde71b1c3b
urllib3==2.2.1
wasabi==1.1.3
weasel==0.3.4
